% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeMaterialPicker.R
\name{updateDatetimeMaterialPickerInput}
\alias{updateDatetimeMaterialPickerInput}
\title{Update a datetime material picker widget}
\usage{
updateDatetimeMaterialPickerInput(session, inputId, value)
}
\arguments{
\item{session}{the Shiny \code{session} object}

\item{inputId}{the id of the datetime material picker widget to be updated}

\item{value}{new value for the datetime material picker widget}
}
\value{
No returned value, this just updates the widget.
}
\description{
Change the value of a datetime material picker input.
}
