% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeSliderPicker.R
\name{updateDatetimeSliderPickerInput}
\alias{updateDatetimeSliderPickerInput}
\title{Update a datetime slider picker widget}
\usage{
updateDatetimeSliderPickerInput(session, inputId, value)
}
\arguments{
\item{session}{the Shiny \code{session} object}

\item{inputId}{the id of the datetime slider picker widget to be updated}

\item{value}{new value for the datetime slider picker widget}
}
\value{
No returned value, this just updates the widget.
}
\description{
Change the value of a datetime slider picker input.
}
