% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeMaterialPicker.R
\name{datetimeMaterialPickerInput}
\alias{datetimeMaterialPickerInput}
\title{<Add Title>}
\usage{
datetimeMaterialPickerInput(inputId, value = NULL, style = NULL)
}
\description{
<Add Description>
}
