% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeSliderPicker.R
\name{datetimeSliderPickerInput}
\alias{datetimeSliderPickerInput}
\title{Datetime picker}
\usage{
datetimeSliderPickerInput(inputId, value = NULL, second = FALSE, save = FALSE)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value}

\item{value}{initial value of the datetime picker; if \code{NULL}, the
initial value is set to the system time}

\item{second}{logical, whether to enable the second picker}

\item{save}{logical, whether to enable the 'save' button}
}
\description{
A datetime picker for a Shiny UI.
}
